+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 ## Changed
 - Nginx is now configured to allow file uploads up to 100MB. ([#126](https://github.com/craftcms/nitro/issues/126))
 - Databases are now backed up automatically when a machine is destroyed. ([#136](https://github.com/craftcms/nitro/issues/136))
-- The OpCache extension is no longer installed by default. ([#129](https://github.com/craftcms/nitro/issues/129))
+- The OPcache extension is no longer installed by default. ([#129](https://github.com/craftcms/nitro/issues/129))
 - When creating a new machine, the `DB_USER` and `DB_PASSWORD` are automatically set in the environment. ([#119](https://github.com/craftcms/nitro/issues/119))
 - The default database is now called `nitro` for MySQL engines, to be consistent with PostgreSQL.
 - The `destroy` command now prompts for confirmation. ([#116](https://github.com/craftcms/nitro/issues/116))